rename controllers in server to be more specific
