+++ server/controllers/utils.js
@@ -13,3 +13,11 @@ export const featureFlags = (ctx, next) => {
   });
   return next();
 };
+
+export const index = (ctx, next) => ctx.render('pages/index', {
+  pageConfig: createPageConfig({inSection: 'index', path: '/'})
+}) && next();
+
+export const progress = (ctx, next) => ctx.render('pages/progress', {
+  pageConfig: createPageConfig({inSection: 'index', path: '/progress'})
+}) && next();
